+++ src/inventory/IListingPanelProps.ts
@@ -1,0 +1,7 @@
+import { InventoryItem } from "../domain/types/InventoryItem";
+
+export interface IListingPanelProps {
+    isOpen: boolean;
+    onDismiss: () => void;
+    item?: InventoryItem;
+}
