+++ src/utils/formatPrice.ts
@@ -1,0 +1,6 @@
+/**
+ * A function to format the price tag for an item.
+ */
+export function formatPrice(price: number) {
+    return `$${price.toFixed(2)}`;
+}
